backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
llama-cpp-python==0.1.50
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2
numpy==1.24.2
scikit-learn==1.2.2
